refactor(landing): LaTeX \frac 정규화 순수 함수를 ts 파일로 분리하고 테스트 추가 (리뷰 #8)
